test: add tests for major nodejs versions
