+++ .github/workflows/test.yml
@@ -9,6 +9,9 @@ on:
 jobs:
   test:
     runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        nodejs: [ "12", "14", "15", "16", "18" ]
 
     steps:
       - uses: actions/checkout@v3
@@ -18,4 +21,11 @@ jobs:
           node-version: "18"
 
       - run: npm ci
+      - run: npm run build
+
+      - name: Use Node.js ${{ matrix.nodejs }}
+        uses: actions/setup-node@v3
+        with:
+          node-version: ${{ matrix.nodejs }}
+
       - run: npm run test
