+++ .gitignore
@@ -1,3 +1,5 @@
 .idea
 node_modules
 lib
+.nyc_output
+coverage
