--- jest.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  moduleFileExtensions: ['ts', 'js'],
-  rootDir: './src',
-  testRegex: '.test.ts$',
-  preset: 'ts-jest',
-  testEnvironment: 'node',
-};
