+++ package.json
@@ -25,7 +25,7 @@
   },
   "scripts": {
     "build": "npm run clean && tsc && rollup -c",
-    "test": "jest",
+    "test": "tap",
     "clean": "rimraf ./lib",
     "prune": "npm run clean && rimraf ./node_modules"
   },
@@ -45,15 +45,14 @@
   "readmeFilename": "README.md",
   "devDependencies": {
     "@rollup/plugin-node-resolve": "^15.1.0",
-    "@types/jest": "^29.5.3",
     "@types/node": "^20.4.4",
+    "@types/tap": "^15.0.8",
     "commitizen": "^4.3.0",
     "cz-conventional-changelog": "^3.3.0",
-    "jest": "^29.6.1",
     "prettier": "^3.0.0",
     "rimraf": "^5.0.1",
     "rollup": "^3.26.3",
-    "ts-jest": "^29.1.1"
+    "tap": "^16.3.7"
   },
   "dependencies": {
     "dotenv": "^16.3.1",
